# Store and reuse field selection data when outputting shortcode options optimization