Using my slow form with 1,0004 fields, the form settings page is pretty slow to load.

It's mostly because of these calls to `FrmField::all_field_selection()`. By calling it only once (instead of nearly 900 times), it cuts my page load by about 69% 🚀.

This optimization also helps on the form builder where we also use this sidebar for populating HTML/Rich Text fields.

_On form settings_
**Before**
~10.43 seconds

**After**
~3.24 seconds

**~7.19 seconds faster**

_In the form builder_
**Before**
~48.3 seconds

**After**
~41.31 seconds

**~6.99 seconds faster**